packages/pykit-grpc/pyproject.toml
[project]
name = "pykit-grpc"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "grpcio",
    "pykit-errors",
    "pykit-component",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pykit_grpc"]
